requirements.txt
# Conexión a la base de datos
psycopg2-binary>=2.9.5

# Web scraping
requests>==2.31.0
beautifulsoup4>==4.12.2
lxml>==4.9.3

# Análisis de datos
pandas>==2.0.0
numpy>==1.24.3
scikit-learn>==1.2.2
joblib>==1.2.0

# Visualización
matplotlib>==3.7.1
tabulate>==0.9.0

# Aplicación web
flask>==2.3.3
jinja2>==3.1.2
gunicorn>==21.2.0

# Utilidades
python-dateutil>==3.12.3